+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/error/WarningsLibrary.java
@@ -64,4 +64,15 @@ public abstract class WarningsLibrary extends Library {
   public Object removeWarnings(Object receiver) throws UnsupportedMessageException {
     throw UnsupportedMessageException.create();
   }
+
+  /**
+   * Checks if the receiver reached a maximal number of warnings that could be reported.
+   *
+   * @param receiver the receiver to analyze
+   * @return whether the receiver reached a maximal number of warnings
+   */
+  @GenerateLibrary.Abstract(ifExported = {"hasWarnings"})
+  public boolean isLimitReached(Object receiver) {
+    return false;
+  }
 }